add overwrite from current comp feature, fixed GroupSearchMode to not overwrite existing preset Jobs
